requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
openai
pydantic-settings
arxiv
